+++ sdk/tables/azure-data-tables/azure/data/tables/_table_service_client.py
@@ -42,11 +42,11 @@ class TableServiceClient(TablesBaseClient):
         The URL to the table service endpoint. Any other entities included
         in the URL path (e.g. table) will be discarded. This URL can be optionally
         authenticated with a SAS token.
-    :param credential:
+    :keyword credential:
         The credentials with which to authenticate. This is optional if the
         account URL already has a SAS token. The value can be one of AzureNamedKeyCredential
         or AzureSasCredential from azure-core.
-    :type credential:
+    :paramtype credential:
         :class:`~azure.core.credentials.AzureNamedKeyCredential` or
         :class:`~azure.core.credentials.AzureSasCredential`
     :keyword str api_version: