+++ sdk/tables/azure-data-tables/tests/test_table_service_properties_async.py
@@ -26,7 +26,7 @@ class TableServicePropertiesTest(AzureTestCase, TableTestCase):
     async def test_table_service_properties_async(self, tables_storage_account_name, tables_primary_storage_account_key):
         # Arrange
         url = self.account_url(tables_storage_account_name, "table")
-        tsc = TableServiceClient(url, tables_primary_storage_account_key, logging_enable=True)
+        tsc = TableServiceClient(url, credential=tables_primary_storage_account_key, logging_enable=True)
         # Act
         resp = await tsc.set_service_properties(
             analytics_logging=TableAnalyticsLogging(),
@@ -45,7 +45,7 @@ class TableServicePropertiesTest(AzureTestCase, TableTestCase):
     async def test_set_logging_async(self, tables_storage_account_name, tables_primary_storage_account_key):
         # Arrange
         url = self.account_url(tables_storage_account_name, "table")
-        tsc = TableServiceClient(url, tables_primary_storage_account_key)
+        tsc = TableServiceClient(url, credential=tables_primary_storage_account_key)
         logging = TableAnalyticsLogging(read=True, write=True, delete=True, retention_policy=TableRetentionPolicy(enabled=True, days=5))
 
         # Act
@@ -61,7 +61,7 @@ class TableServicePropertiesTest(AzureTestCase, TableTestCase):
     async def test_set_hour_metrics_async(self, tables_storage_account_name, tables_primary_storage_account_key):
         # Arrange
         url = self.account_url(tables_storage_account_name, "table")
-        tsc = TableServiceClient(url, tables_primary_storage_account_key)
+        tsc = TableServiceClient(url, credential=tables_primary_storage_account_key)
         hour_metrics = TableMetrics(enabled=True, include_apis=True, retention_policy=TableRetentionPolicy(enabled=True, days=5))
 
         # Act
@@ -77,7 +77,7 @@ class TableServicePropertiesTest(AzureTestCase, TableTestCase):
     async def test_set_minute_metrics_async(self, tables_storage_account_name, tables_primary_storage_account_key):
         # Arrange
         url = self.account_url(tables_storage_account_name, "table")
-        tsc = TableServiceClient(url, tables_primary_storage_account_key)
+        tsc = TableServiceClient(url, credential=tables_primary_storage_account_key)
         minute_metrics = TableMetrics(enabled=True, include_apis=True,
                                  retention_policy=TableRetentionPolicy(enabled=True, days=5))
 
@@ -94,7 +94,7 @@ class TableServicePropertiesTest(AzureTestCase, TableTestCase):
     async def test_set_cors_async(self, tables_storage_account_name, tables_primary_storage_account_key):
         # Arrange
         url = self.account_url(tables_storage_account_name, "table")
-        tsc = TableServiceClient(url, tables_primary_storage_account_key)
+        tsc = TableServiceClient(url, credential=tables_primary_storage_account_key)
         cors_rule1 = TableCorsRule(['www.xyz.com'], ['GET'])
 
         allowed_origins = ['www.xyz.com', "www.ab.com", "www.bc.com"]
@@ -122,7 +122,7 @@ class TableServicePropertiesTest(AzureTestCase, TableTestCase):
     @tables_decorator_async
     async def test_too_many_cors_rules_async(self, tables_storage_account_name, tables_primary_storage_account_key):
         # Arrange
-        tsc = TableServiceClient(self.account_url(tables_storage_account_name, "table"), tables_primary_storage_account_key)
+        tsc = TableServiceClient(self.account_url(tables_storage_account_name, "table"), credential=tables_primary_storage_account_key)
         cors = []
         for i in range(0, 6):
             cors.append(TableCorsRule(['www.xyz.com'], ['GET']))
@@ -134,7 +134,7 @@ class TableServicePropertiesTest(AzureTestCase, TableTestCase):
     @tables_decorator_async
     async def test_retention_too_long_async(self, tables_storage_account_name, tables_primary_storage_account_key):
         # Arrange
-        tsc = TableServiceClient(self.account_url(tables_storage_account_name, "table"), tables_primary_storage_account_key)
+        tsc = TableServiceClient(self.account_url(tables_storage_account_name, "table"), credential=tables_primary_storage_account_key)
         minute_metrics = TableMetrics(enabled=True, include_apis=True,
                                  retention_policy=TableRetentionPolicy(enabled=True, days=366))